Modal/pop-up card on homepage (fixed)
A modal appears to show additional information when Apply button is pressed on the homepage. Currently it just repeats the description, but it should be expanded upon. These changes should also be implemented in the search page next, and search page changes can be brought to the homepage.

Now fixed for bcrypt.