Added clipboard functionality to jobsearch
There's a clipboard button on each job modal on the jobsearch page which copies the current job listing url. Opening the url redirects to the same search, same page and same job modal. Also fixed bugs with the url system.

I changed the icon library in the head so the about-us page has missing icons - need to fix later.