UI optimised for mobile & other UI improvements
Multiple adjustments to elements so they now scale correctly or hide on mobile and small devices.
Added share API button to replace the copy to clipboard button for mobile devices.
Added clipboard/share buttons on homepage (not fully functional as there's no URL generator).
Added a loading animation on the homepage and searchpage.