User is now able to create and post a featured listing.
Mostly functional, a user can create and "buy" a custom featured listing and it will appear on the home page. The Stripe payment gateway is still in test mode, the keys will have to be updated and secured in the .env on release (when JobSpy is a registered company). Next I will need to add feedback when a payment has failed or succeeded. I should also add the ability to edit/renew listings. And I should make featured listings appear on search results.