Create listing (fixed)
Mostly functional, a user can create and "buy" a custom featured listing and it will appear on the home page (and the listings expire after a month). The Stripe payment gateway is still in test mode, the keys will have to be updated and secured in the .env on release (when JobSpy is a registered company). Currently you must load up the Stripe CLI to simulate a successful payment (https://stripe.com/docs/payments/checkout/fulfill-orders#install-stripe-cli). Next I will need to add feedback when a payment has failed or succeeded. I should also add the ability to edit/renew listings. And I should make featured listings appear on search results.