Bug fixes and UI Unification
Some UI tweaks. Most notably the UI between the home page and the search page have been made to look and behave the same, with a refreshed look for job posts (now the average salaries are contained inside the modal).
Additionally some bug fixes including a major issue with pages, the social media links not visible in the about us page,